refactor(rust): collapse exec-control codec layer
