+++ crates/vsock-proto/src/payloads/mod.rs
@@ -1,6 +1,5 @@
 pub(crate) mod empty;
 pub(crate) mod error;
-pub(crate) mod exec_control;
 pub(crate) mod exec_operation;
 pub(crate) mod write_file;
